refactor(api): break monolithic database_operations into specialized modules
Transformed the 727-line database_operations.py monolith into a focused
ensemble cast of specialized modules. Each module now has a single
responsibility, making MCP tools happy and code maintenance a joy.

The new cast includes:
- db_connection.py (80 lines): Database connection choreographer
- db_script_ops.py (210 lines): Script operations virtuoso
- db_scene_ops.py (253 lines): Scene and content maestro
- db_embedding_ops.py (221 lines): Embedding storage specialist
- database_operations.py (305 lines): The facade director

Full backward compatibility maintained - existing APIs unchanged.
All tests pass without modification. This is modular design done right.

"I'm gonna make him an offer he can't refuse." - Don Vito Corleone, The Godfather (1972)
