+++ src/scriptrag/api/database_operations.py
@@ -1,35 +1,33 @@
-"""Database operations for indexing scripts into ScriptRAG database."""
+"""Database operations for indexing scripts into ScriptRAG database.
 
-import json
+This module serves as the main interface for database operations, delegating
+to specialized modules for connection management, script operations, scene
+operations, and embedding operations.
+"""
+
 import sqlite3
 from collections.abc import Generator
 from contextlib import contextmanager
-from dataclasses import dataclass
-from datetime import datetime
 from pathlib import Path
 from typing import Any
 
-from scriptrag.config import ScriptRAGSettings, get_logger
-from scriptrag.exceptions import DatabaseError
+from scriptrag.api.db_connection import DatabaseConnectionManager
+from scriptrag.api.db_embedding_ops import EmbeddingOperations
+from scriptrag.api.db_scene_ops import SceneOperations
+from scriptrag.api.db_script_ops import ScriptOperations, ScriptRecord
+from scriptrag.config import ScriptRAGSettings
 from scriptrag.parser import Dialogue, Scene, Script
-from scriptrag.utils import ScreenplayUtils
 
-logger = get_logger(__name__)
-
-
-@dataclass
-class ScriptRecord:
-    """Database record for a script."""
-
-    id: int | None = None
-    title: str | None = None
-    author: str | None = None
-    file_path: str | None = None
-    metadata: dict[str, Any] | None = None
+# Re-export ScriptRecord for backward compatibility
+__all__ = ["DatabaseOperations", "ScriptRecord"]
 
 
 class DatabaseOperations:
-    """Handles all database operations for indexing."""
+    """Handles all database operations for indexing.
+
+    This class coordinates between specialized operation modules to provide
+    a unified interface for database operations.
+    """
 
     def __init__(self, settings: ScriptRAGSettings) -> None:
         """Initialize database operations with settings.
@@ -40,29 +38,20 @@ class DatabaseOperations:
         self.settings = settings
         self.db_path = settings.database_path
 
+        # Initialize specialized operation handlers
+        self._conn_manager = DatabaseConnectionManager(settings)
+        self._script_ops = ScriptOperations()
+        self._scene_ops = SceneOperations()
+        self._embedding_ops = EmbeddingOperations()
+
+    # Connection management - delegate to connection manager
     def get_connection(self) -> sqlite3.Connection:
         """Get database connection with proper configuration.
 
         Returns:
             Configured SQLite connection
         """
-        conn = sqlite3.connect(
-            str(self.db_path), timeout=self.settings.database_timeout
-        )
-
-        # Configure connection
-        conn.execute(f"PRAGMA journal_mode = {self.settings.database_journal_mode}")
-        conn.execute(f"PRAGMA synchronous = {self.settings.database_synchronous}")
-        conn.execute(f"PRAGMA cache_size = {self.settings.database_cache_size}")
-        conn.execute(f"PRAGMA temp_store = {self.settings.database_temp_store}")
-
-        if self.settings.database_foreign_keys:
-            conn.execute("PRAGMA foreign_keys = ON")
-
-        # Enable JSON support
-        conn.row_factory = sqlite3.Row
-
-        return conn
+        return self._conn_manager.get_connection()
 
     @contextmanager
     def transaction(self) -> Generator[sqlite3.Connection, None, None]:
@@ -71,15 +60,8 @@ class DatabaseOperations:
         Yields:
             Database connection within a transaction context
         """
-        conn = self.get_connection()
-        try:
+        with self._conn_manager.transaction() as conn:
             yield conn
-            conn.commit()
-        except Exception:
-            conn.rollback()
-            raise
-        finally:
-            conn.close()
 
     def check_database_exists(self) -> bool:
         """Check if the database exists and is initialized.
@@ -87,19 +69,9 @@ class DatabaseOperations:
         Returns:
             True if database exists and has schema, False otherwise
         """
-        if not self.db_path.exists():
-            return False
+        return self._conn_manager.check_database_exists()
 
-        try:
-            with self.transaction() as conn:
-                cursor = conn.execute(
-                    "SELECT name FROM sqlite_master "
-                    "WHERE type='table' AND name='scripts'"
-                )
-                return cursor.fetchone() is not None
-        except sqlite3.Error:
-            return False
-
+    # Script operations - delegate to script operations module
     def get_existing_script(
         self, conn: sqlite3.Connection, file_path: Path
     ) -> ScriptRecord | None:
@@ -112,22 +84,7 @@ class DatabaseOperations:
         Returns:
             ScriptRecord if found, None otherwise
         """
-        cursor = conn.execute(
-            "SELECT id, title, author, file_path, metadata "
-            "FROM scripts WHERE file_path = ?",
-            (str(file_path),),
-        )
-        row = cursor.fetchone()
-
-        if row:
-            return ScriptRecord(
-                id=row["id"],
-                title=row["title"],
-                author=row["author"],
-                file_path=row["file_path"],
-                metadata=json.loads(row["metadata"]) if row["metadata"] else None,
-            )
-        return None
+        return self._script_ops.get_existing_script(conn, file_path)
 
     def upsert_script(
         self, conn: sqlite3.Connection, script: Script, file_path: Path
@@ -142,82 +99,7 @@ class DatabaseOperations:
         Returns:
             ID of the inserted or updated script
         """
-        metadata = script.metadata.copy() if script.metadata else {}
-        metadata["last_indexed"] = datetime.now().isoformat()
-
-        # Extract series/episode info from metadata if available
-        # Ensure we have safe defaults for all fields
-        title = script.title or "Untitled"
-        author = script.author or "Unknown"
-        project_title = metadata.get("project_title") or title
-        series_title = metadata.get("series_title")
-        season = metadata.get("season")
-        episode = metadata.get("episode")
-
-        # Use separate SELECT/UPDATE/INSERT for compatibility
-        # This works with older SQLite versions that don't support ON CONFLICT RETURNING
-        cursor = conn.execute(
-            "SELECT id FROM scripts WHERE file_path = ?", (str(file_path),)
-        )
-        existing = cursor.fetchone()
-
-        if existing:
-            # Update existing script
-            conn.execute(
-                """
-                UPDATE scripts SET
-                    title = ?, author = ?, project_title = ?,
-                    series_title = ?, season = ?, episode = ?,
-                    metadata = ?, updated_at = CURRENT_TIMESTAMP
-                WHERE file_path = ?
-                """,
-                (
-                    title,
-                    author,
-                    project_title,
-                    series_title,
-                    season,
-                    episode,
-                    json.dumps(metadata),
-                    str(file_path),
-                ),
-            )
-            script_id = existing[0]
-            logger.debug(f"Updated script {script_id}: {title} at {file_path}")
-        else:
-            # Insert new script
-            cursor = conn.execute(
-                """
-                INSERT INTO scripts (
-                    file_path, title, author, project_title,
-                    series_title, season, episode, metadata
-                ) VALUES (?, ?, ?, ?, ?, ?, ?, ?)
-                """,
-                (
-                    str(file_path),
-                    title,
-                    author,
-                    project_title,
-                    series_title,
-                    season,
-                    episode,
-                    json.dumps(metadata),
-                ),
-            )
-            script_id = cursor.lastrowid
-            if script_id is None:
-                raise DatabaseError(
-                    message="Failed to get script ID after insert",
-                    hint="Database constraint violation or transaction issue",
-                    details={
-                        "script_title": title,
-                        "script_path": str(file_path),
-                        "operation": "INSERT INTO scripts",
-                    },
-                )
-            logger.debug(f"Inserted script {script_id}: {title} at {file_path}")
-
-        return int(script_id)
+        return self._script_ops.upsert_script(conn, script, file_path)
 
     def clear_script_data(self, conn: sqlite3.Connection, script_id: int) -> None:
         """Clear all existing data for a script before re-indexing.
@@ -226,11 +108,23 @@ class DatabaseOperations:
             conn: Database connection
             script_id: ID of the script to clear
         """
-        # Delete all related data (cascades will handle most)
-        conn.execute("DELETE FROM scenes WHERE script_id = ?", (script_id,))
-        conn.execute("DELETE FROM characters WHERE script_id = ?", (script_id,))
-        logger.debug(f"Cleared existing data for script {script_id}")
+        self._script_ops.clear_script_data(conn, script_id)
 
+    def get_script_stats(
+        self, conn: sqlite3.Connection, script_id: int
+    ) -> dict[str, int]:
+        """Get statistics for an indexed script.
+
+        Args:
+            conn: Database connection
+            script_id: ID of the script
+
+        Returns:
+            Dictionary with counts of scenes, characters, dialogues, and actions
+        """
+        return self._script_ops.get_script_stats(conn, script_id)
+
+    # Scene operations - delegate to scene operations module
     def upsert_scene(
         self, conn: sqlite3.Connection, scene: Scene, script_id: int
     ) -> tuple[int, bool]:
@@ -245,98 +139,7 @@ class DatabaseOperations:
             Tuple of (scene_id, content_changed) where content_changed is True
             if the scene content has changed
         """
-        # Extract location and time from heading
-        location = (
-            scene.location
-            if scene.location
-            else ScreenplayUtils.extract_location(scene.heading)
-        )
-        time_of_day = (
-            scene.time_of_day
-            if scene.time_of_day
-            else ScreenplayUtils.extract_time(scene.heading)
-        )
-
-        # Prepare metadata
-        metadata: dict[str, Any] = {}
-        if scene.boneyard_metadata:
-            metadata["boneyard"] = scene.boneyard_metadata
-        metadata["content_hash"] = scene.content_hash
-
-        # Check if scene exists
-        cursor = conn.execute(
-            "SELECT id, metadata FROM scenes WHERE script_id = ? AND scene_number = ?",
-            (script_id, scene.number),
-        )
-        existing = cursor.fetchone()
-
-        scene_id: int
-        content_changed = False
-
-        if existing:
-            # Check if content has actually changed by comparing hashes
-            existing_metadata = (
-                json.loads(existing["metadata"]) if existing["metadata"] else {}
-            )
-            existing_hash = existing_metadata.get("content_hash")
-            content_changed = existing_hash != scene.content_hash
-
-            # Update existing scene
-            conn.execute(
-                """
-                UPDATE scenes
-                SET heading = ?, location = ?, time_of_day = ?, content = ?,
-                    metadata = ?, updated_at = CURRENT_TIMESTAMP
-                WHERE id = ?
-                """,
-                (
-                    scene.heading,
-                    location,
-                    time_of_day,
-                    scene.content,
-                    json.dumps(metadata),
-                    existing["id"],
-                ),
-            )
-            scene_id = int(existing["id"])
-            logger.debug(
-                f"Updated scene {scene_id}: {scene.heading}, "
-                f"content_changed={content_changed}"
-            )
-        else:
-            # Insert new scene
-            cursor = conn.execute(
-                """
-                INSERT INTO scenes (script_id, scene_number, heading, location,
-                                  time_of_day, content, metadata)
-                VALUES (?, ?, ?, ?, ?, ?, ?)
-                """,
-                (
-                    script_id,
-                    scene.number,
-                    scene.heading,
-                    location,
-                    time_of_day,
-                    scene.content,
-                    json.dumps(metadata),
-                ),
-            )
-            lastrowid = cursor.lastrowid
-            if lastrowid is None:
-                raise DatabaseError(
-                    message="Failed to get scene ID after database insert",
-                    hint="Database constraint violation or transaction issue",
-                    details={
-                        "scene_number": getattr(scene, "scene_number", "unknown"),
-                        "script_id": script_id,
-                        "operation": "INSERT INTO scenes",
-                    },
-                )
-            scene_id = lastrowid
-            content_changed = True  # New scene, so content has "changed"
-            logger.debug(f"Inserted scene {scene_id}: {scene.heading}")
-
-        return scene_id, content_changed
+        return self._scene_ops.upsert_scene(conn, scene, script_id)
 
     def upsert_characters(
         self, conn: sqlite3.Connection, script_id: int, characters: set[str]
@@ -351,31 +154,7 @@ class DatabaseOperations:
         Returns:
             Mapping of character names to their IDs
         """
-        character_map = {}
-
-        for name in characters:
-            # Check if character exists
-            cursor = conn.execute(
-                "SELECT id FROM characters WHERE script_id = ? AND name = ?",
-                (script_id, name),
-            )
-            existing = cursor.fetchone()
-
-            if existing:
-                character_map[name] = existing["id"]
-            else:
-                # Insert new character
-                cursor = conn.execute(
-                    """
-                    INSERT INTO characters (script_id, name)
-                    VALUES (?, ?)
-                    """,
-                    (script_id, name),
-                )
-                character_map[name] = cursor.lastrowid
-                logger.debug(f"Inserted character {character_map[name]}: {name}")
-
-        return character_map
+        return self._scene_ops.upsert_characters(conn, script_id, characters)
 
     def clear_scene_content(self, conn: sqlite3.Connection, scene_id: int) -> None:
         """Clear dialogues and actions for a scene before re-inserting.
@@ -384,8 +163,7 @@ class DatabaseOperations:
             conn: Database connection
             scene_id: ID of the scene to clear
         """
-        conn.execute("DELETE FROM dialogues WHERE scene_id = ?", (scene_id,))
-        conn.execute("DELETE FROM actions WHERE scene_id = ?", (scene_id,))
+        self._scene_ops.clear_scene_content(conn, scene_id)
 
     def insert_dialogues(
         self,
@@ -405,38 +183,9 @@ class DatabaseOperations:
         Returns:
             Number of dialogues inserted
         """
-        count = 0
-        for order, dialogue in enumerate(dialogues):
-            if dialogue.character not in character_map:
-                # Skip dialogues for unknown characters - this can happen when a
-                # dialogue is attributed to a character that wasn't properly
-                # extracted during character discovery phase, possibly due to
-                # formatting issues in the screenplay
-                logger.warning(f"Unknown character in dialogue: {dialogue.character}")
-                continue
-
-            metadata = {}
-            if dialogue.parenthetical:
-                metadata["parenthetical"] = dialogue.parenthetical
-
-            conn.execute(
-                """
-                INSERT INTO dialogues (scene_id, character_id, dialogue_text,
-                                     order_in_scene, metadata)
-                VALUES (?, ?, ?, ?, ?)
-                """,
-                (
-                    scene_id,
-                    character_map[dialogue.character],
-                    dialogue.text,
-                    order,
-                    json.dumps(metadata) if metadata else None,
-                ),
-            )
-            count += 1
-
-        logger.debug(f"Inserted {count} dialogues for scene {scene_id}")
-        return count
+        return self._scene_ops.insert_dialogues(
+            conn, scene_id, dialogues, character_map
+        )
 
     def insert_actions(
         self, conn: sqlite3.Connection, scene_id: int, actions: list[str]
@@ -451,73 +200,9 @@ class DatabaseOperations:
         Returns:
             Number of actions inserted
         """
-        count = 0
-        for order, action_text in enumerate(actions):
-            if not action_text.strip():
-                continue
+        return self._scene_ops.insert_actions(conn, scene_id, actions)
 
-            conn.execute(
-                """
-                INSERT INTO actions (scene_id, action_text, order_in_scene)
-                VALUES (?, ?, ?)
-                """,
-                (scene_id, action_text, order),
-            )
-            count += 1
-
-        logger.debug(f"Inserted {count} actions for scene {scene_id}")
-        return count
-
-    def get_script_stats(
-        self, conn: sqlite3.Connection, script_id: int
-    ) -> dict[str, int]:
-        """Get statistics for an indexed script.
-
-        Args:
-            conn: Database connection
-            script_id: ID of the script
-
-        Returns:
-            Dictionary with counts of scenes, characters, dialogues, and actions
-        """
-        stats = {}
-
-        # Count scenes
-        cursor = conn.execute(
-            "SELECT COUNT(*) as count FROM scenes WHERE script_id = ?", (script_id,)
-        )
-        stats["scenes"] = cursor.fetchone()["count"]
-
-        # Count characters
-        cursor = conn.execute(
-            "SELECT COUNT(*) as count FROM characters WHERE script_id = ?", (script_id,)
-        )
-        stats["characters"] = cursor.fetchone()["count"]
-
-        # Count dialogues
-        cursor = conn.execute(
-            """
-            SELECT COUNT(*) as count FROM dialogues d
-            JOIN scenes s ON d.scene_id = s.id
-            WHERE s.script_id = ?
-            """,
-            (script_id,),
-        )
-        stats["dialogues"] = cursor.fetchone()["count"]
-
-        # Count actions
-        cursor = conn.execute(
-            """
-            SELECT COUNT(*) as count FROM actions a
-            JOIN scenes s ON a.scene_id = s.id
-            WHERE s.script_id = ?
-            """,
-            (script_id,),
-        )
-        stats["actions"] = cursor.fetchone()["count"]
-
-        return stats
-
+    # Embedding operations - delegate to embedding operations module
     def upsert_embedding(
         self,
         conn: sqlite3.Connection,
@@ -540,65 +225,14 @@ class DatabaseOperations:
         Returns:
             ID of the inserted or updated embedding
         """
-        # Note: embedding_path is reserved for future Git LFS storage support
-        # Currently, we store binary data directly in the database
-        _ = embedding_path  # Mark as intentionally unused
-
-        # Check if embedding exists
-        cursor = conn.execute(
-            """
-            SELECT id FROM embeddings
-            WHERE entity_type = ? AND entity_id = ? AND embedding_model = ?
-            """,
-            (entity_type, entity_id, embedding_model),
+        return self._embedding_ops.upsert_embedding(
+            conn,
+            entity_type,
+            entity_id,
+            embedding_model,
+            embedding_data,
+            embedding_path,
         )
-        existing = cursor.fetchone()
-
-        # For Git LFS storage, we store the path reference instead of raw data
-        # The actual embedding will be loaded from the file when needed
-        embedding_blob = embedding_data if embedding_data else b""
-
-        if existing:
-            # Update existing embedding
-            conn.execute(
-                """
-                UPDATE embeddings
-                SET embedding = ?, updated_at = CURRENT_TIMESTAMP
-                WHERE id = ?
-                """,
-                (embedding_blob, existing["id"]),
-            )
-            embedding_id = int(existing["id"])
-            logger.debug(
-                f"Updated embedding {embedding_id} for {entity_type}:{entity_id}"
-            )
-        else:
-            # Insert new embedding
-            cursor = conn.execute(
-                """
-                INSERT INTO embeddings
-                (entity_type, entity_id, embedding_model, embedding)
-                VALUES (?, ?, ?, ?)
-                """,
-                (entity_type, entity_id, embedding_model, embedding_blob),
-            )
-            lastrowid = cursor.lastrowid
-            if lastrowid is None:
-                raise DatabaseError(
-                    message="Failed to get embedding ID after database insert",
-                    hint="Database constraint violation or embedding storage issue",
-                    details={
-                        "entity_type": entity_type,
-                        "entity_id": entity_id,
-                        "operation": "INSERT INTO embeddings",
-                    },
-                )
-            embedding_id = lastrowid
-            logger.debug(
-                f"Inserted embedding {embedding_id} for {entity_type}:{entity_id}"
-            )
-
-        return embedding_id
 
     def get_scene_embeddings(
         self,
@@ -616,24 +250,9 @@ class DatabaseOperations:
         Returns:
             List of (scene_id, embedding_data) tuples
         """
-        cursor = conn.execute(
-            """
-            SELECT s.id, e.embedding
-            FROM scenes s
-            JOIN embeddings e ON e.entity_id = s.id
-            WHERE s.script_id = ?
-            AND e.entity_type = 'scene'
-            AND e.embedding_model = ?
-            AND e.embedding IS NOT NULL
-            """,
-            (script_id, embedding_model),
+        return self._embedding_ops.get_scene_embeddings(
+            conn, script_id, embedding_model
         )
-
-        results = []
-        for row in cursor:
-            results.append((row["id"], row["embedding"]))
-
-        return results
 
     def get_embedding(
         self,
@@ -653,20 +272,14 @@ class DatabaseOperations:
         Returns:
             Binary embedding data if found, None otherwise
         """
-        cursor = conn.execute(
-            """
-            SELECT embedding FROM embeddings
-            WHERE entity_type = ? AND entity_id = ? AND embedding_model = ?
-            """,
-            (entity_type, entity_id, embedding_model),
+        return self._embedding_ops.get_embedding(
+            conn, entity_type, entity_id, embedding_model
         )
-        row = cursor.fetchone()
-        return row["embedding"] if row else None
 
     def search_similar_scenes(
         self,
         conn: sqlite3.Connection,
-        query_embedding: bytes,  # noqa: ARG002
+        query_embedding: bytes,
         script_id: int | None,
         embedding_model: str,
         limit: int = 10,
@@ -687,41 +300,6 @@ class DatabaseOperations:
         Returns:
             List of scene records with similarity scores
         """
-        # Build query based on whether script_id is provided
-        params: tuple[Any, ...]
-        if script_id:
-            query = """
-                SELECT s.*, e.embedding
-                FROM scenes s
-                JOIN embeddings e ON e.entity_id = s.id
-                WHERE s.script_id = ?
-                AND e.entity_type = 'scene'
-                AND e.embedding_model = ?
-                AND e.embedding IS NOT NULL
-            """
-            params = (script_id, embedding_model)
-        else:
-            query = """
-                SELECT s.*, e.embedding
-                FROM scenes s
-                JOIN embeddings e ON e.entity_id = s.id
-                WHERE e.entity_type = 'scene'
-                AND e.embedding_model = ?
-                AND e.embedding IS NOT NULL
-            """
-            params = (embedding_model,)
-
-        cursor = conn.execute(query, params)
-
-        # Note: For production use with large datasets, the similarity
-        # calculation should be done in a specialized vector database
-        # or using database extensions like sqlite-vss
-        scenes = []
-        for row in cursor:
-            scene_dict = dict(row)
-            # Store embedding for similarity calculation
-            scene_dict["_embedding"] = scene_dict.pop("embedding")
-            scenes.append(scene_dict)
-
-        # Return scenes with embeddings for external similarity calculation
-        return scenes[:limit]  # Limit applied after retrieval for now
+        return self._embedding_ops.search_similar_scenes(
+            conn, query_embedding, script_id, embedding_model, limit
+        )
